Move events inside commons package
