+++ src/main/java/seedu/address/MainApp.java
@@ -7,8 +7,8 @@ import javafx.stage.Stage;
 import seedu.address.commons.core.Version;
 import seedu.address.model.AddressBook;
 import seedu.address.ui.UiManager;
-import seedu.address.events.EventManager;
-import seedu.address.events.controller.ExitAppRequestEvent;
+import seedu.address.commons.core.EventManager;
+import seedu.address.commons.events.controller.ExitAppRequestEvent;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.storage.StorageManager;
