+++ src/main/java/seedu/address/commons/events/BaseEvent.java
@@ -1,6 +1,6 @@
-package seedu.address.events;
+package seedu.address.commons.events;
 
-import seedu.address.events.controller.JumpToListRequestEvent;
+import seedu.address.commons.events.controller.JumpToListRequestEvent;
 
 public abstract class BaseEvent {
 
@@ -17,7 +17,7 @@ public abstract class BaseEvent {
      *     even if they are two different event objects. This is a weaker form of equality.
      *     e.g. if both events intended the selection to jump to the 1st element in the list. <br>
      *     The default implementation returns true of both objects are of the same class.
-     *     For example, comparing two distinct {@link seedu.address.events.controller.ExitAppRequestEvent} will return
+     *     For example, comparing two distinct {@link seedu.address.commons.events.controller.ExitAppRequestEvent} will return
      *     true (because both intend to exit the app). <br>
      *     If the events are parameterized, e.g. {@link JumpToListRequestEvent} this method
      *     should be overriden to compare the parameters.
