+++ src/main/java/seedu/address/storage/StorageManager.java
@@ -2,8 +2,8 @@ package seedu.address.storage;
 
 import com.google.common.eventbus.Subscribe;
 import seedu.address.commons.util.FileUtil;
-import seedu.address.events.model.LocalModelChangedEvent;
-import seedu.address.events.storage.*;
+import seedu.address.commons.events.model.LocalModelChangedEvent;
+import seedu.address.commons.events.storage.*;
 import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.core.ComponentManager;
 import seedu.address.model.UserPrefs;
