+++ src/main/java/seedu/address/ui/UiManager.java
@@ -12,11 +12,11 @@ import seedu.address.commons.core.ComponentManager;
 import seedu.address.commons.core.Config;
 import seedu.address.commons.core.LoggerManager;
 import seedu.address.commons.util.StringUtil;
-import seedu.address.events.controller.JumpToListRequestEvent;
-import seedu.address.events.controller.PersonPanelSelectionChangedEvent;
-import seedu.address.events.controller.ShowHelpEvent;
-import seedu.address.events.storage.FileOpeningExceptionEvent;
-import seedu.address.events.storage.FileSavingExceptionEvent;
+import seedu.address.commons.events.controller.JumpToListRequestEvent;
+import seedu.address.commons.events.controller.PersonPanelSelectionChangedEvent;
+import seedu.address.commons.events.controller.ShowHelpEvent;
+import seedu.address.commons.events.storage.FileOpeningExceptionEvent;
+import seedu.address.commons.events.storage.FileSavingExceptionEvent;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.ReadOnlyPerson;
