+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -7,10 +7,10 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import seedu.address.logic.commands.*;
-import seedu.address.events.EventManager;
-import seedu.address.events.controller.JumpToListRequestEvent;
-import seedu.address.events.controller.ShowHelpEvent;
-import seedu.address.events.model.LocalModelChangedEvent;
+import seedu.address.commons.core.EventManager;
+import seedu.address.commons.events.controller.JumpToListRequestEvent;
+import seedu.address.commons.events.controller.ShowHelpEvent;
+import seedu.address.commons.events.model.LocalModelChangedEvent;
 import seedu.address.model.AddressBook;
 import seedu.address.model.ModelManager;
 import seedu.address.model.ReadOnlyAddressBook;
