+++ src/test/java/seedu/address/storage/StorageManagerTest.java
@@ -6,8 +6,8 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import seedu.address.commons.util.FileUtil;
-import seedu.address.events.model.LocalModelChangedEvent;
-import seedu.address.events.storage.SaveDataRequestEvent;
+import seedu.address.commons.events.model.LocalModelChangedEvent;
+import seedu.address.commons.events.storage.SaveDataRequestEvent;
 import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.exceptions.DuplicateTagException;
 import seedu.address.model.ModelManager;
